Suggestor clean
